Formats for Windows

+++ xl2/tao/widget.cpp
@@ -1826,8 +1826,8 @@ ulonglong Widget::elapsed(ulonglong since, ulonglong until,
     {
         char buffer[80];
         snprintf(buffer, sizeof(buffer),
-                 "Duration=%llu-%llu (~%f) %5.2f-%5.2f FPS (~%5.2f)",
-                 tmin, tmax, double(tsum )/ tcount,
+                 "Duration=%lu-%lu (~%f) %5.2f-%5.2f FPS (~%5.2f)",
+                 (ulong) tmin, (ulong) tmax, double(tsum )/ tcount,
                  (100000000ULL / tmax)*0.01,
                  (100000000ULL / tmin)*0.01,
                  (100000000ULL * tcount / tsum) * 0.01);
